+++ src/RuleSet/Php53.php
@@ -810,6 +810,11 @@ final class Php53 extends AbstractRuleSet implements ExplicitRuleSet
         ],
     ];
 
+    public function name(): string
+    {
+        return 'ergebnis (PHP 5.3)';
+    }
+
     public function targetPhpVersion(): int
     {
         return 50300;