+++ src/RuleSet/Php54.php
@@ -812,6 +812,11 @@ final class Php54 extends AbstractRuleSet implements ExplicitRuleSet
         ],
     ];
 
+    public function name(): string
+    {
+        return 'ergebnis (PHP 5.4)';
+    }
+
     public function targetPhpVersion(): int
     {
         return 50400;