+++ src/RuleSet/Php55.php
@@ -821,6 +821,11 @@ final class Php55 extends AbstractRuleSet implements ExplicitRuleSet
         ],
     ];
 
+    public function name(): string
+    {
+        return 'ergebnis (PHP 5.5)';
+    }
+
     public function targetPhpVersion(): int
     {
         return 50500;