+++ src/RuleSet/Php56.php
@@ -821,6 +821,11 @@ final class Php56 extends AbstractRuleSet implements ExplicitRuleSet
         ],
     ];
 
+    public function name(): string
+    {
+        return 'ergebnis (PHP 5.6)';
+    }
+
     public function targetPhpVersion(): int
     {
         return 50600;